+++ TabLayout/src/main/java/com/angcyo/tablayout/DslTabLayout.kt
@@ -120,15 +120,21 @@ open class DslTabLayout(
             onStyleItemView = { itemView, index, select ->
                 tabLayoutConfig?.onStyleItemView?.invoke(itemView, index, select)
             }
-            onSelectItemView = { itemView, index, select ->
-                tabLayoutConfig?.onSelectItemView?.invoke(itemView, index, select) ?: false
+            onSelectItemView = { itemView, index, select, fromUser ->
+                tabLayoutConfig?.onSelectItemView?.invoke(itemView, index, select, fromUser)
+                    ?: false
             }
-            onSelectViewChange = { fromView, selectViewList, reselect ->
-                tabLayoutConfig?.onSelectViewChange?.invoke(fromView, selectViewList, reselect)
+            onSelectViewChange = { fromView, selectViewList, reselect, fromUser ->
+                tabLayoutConfig?.onSelectViewChange?.invoke(
+                    fromView,
+                    selectViewList,
+                    reselect,
+                    fromUser
+                )
             }
-            onSelectIndexChange = { fromIndex, selectList, reselect ->
+            onSelectIndexChange = { fromIndex, selectList, reselect, fromUser ->
                 if (tabLayoutConfig == null) {
-                    "选择:[$fromIndex]->${selectList} reselect:$reselect".logi()
+                    "选择:[$fromIndex]->${selectList} reselect:$reselect fromUser:$fromUser".logi()
                 }
 
                 val toIndex = selectList.last()
@@ -137,8 +143,12 @@ open class DslTabLayout(
                 _scrollToCenter(toIndex, tabIndicator.indicatorAnim)
                 postInvalidate()
 
-                tabLayoutConfig?.onSelectIndexChange?.invoke(fromIndex, selectList, reselect)
-                    ?: _viewPagerDelegate?.onSetCurrentItem(fromIndex, toIndex)
+                tabLayoutConfig?.onSelectIndexChange?.invoke(
+                    fromIndex,
+                    selectList,
+                    reselect,
+                    fromUser
+                ) ?: _viewPagerDelegate?.onSetCurrentItem(fromIndex, toIndex)
             }
         }
     }
@@ -214,6 +224,7 @@ open class DslTabLayout(
     /**设置tab的位置*/
     fun setCurrentItem(index: Int, notify: Boolean = true) {
         if (currentItemIndex == index) {
+            _scrollToCenter(index, tabIndicator.indicatorAnim)
             return
         }
         dslSelector.selector(index, true, notify)
@@ -928,7 +939,12 @@ open class DslTabLayout(
             return
         }
 
-        tabIndicator.currentIndex = max(0, fromIndex)
+        if (fromIndex < 0) {
+            //从一个不存在的位置, 到目标位置
+            tabIndicator.currentIndex = toIndex
+        } else {
+            tabIndicator.currentIndex = fromIndex
+        }
         tabIndicator._targetIndex = toIndex
 
         if (isInEditMode) {